pyproject.toml
[tool.poetry]
name = "realitydb"
version = "0.0.1"
description = "RPC, WebSocket Server, JSON RPC 2.0, OpenAPI, Swagger UI"
authors = ["Oscar Bahamonde <oscar@perudevlabs.com>"]
readme = "README.md"
license = "MIT"
keywords = ["RPC", "WebSocket", "FastAPI", "OpenAPI", "Swagger", "JSON-RPC"]
repository = "https://github.com/perudevlabs/realitydb"
homepage = "https://perudevlabs.com"
documentation = "https://realitydb-docs.perudevlabs.com"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.31.1"}
fastapi = {extras = ["all"], version = "^0.115.2"}
rocksdict = "^0.3.23"
pydantic = "^2.9.2"
orjson = "^3.10.7"
cachetools = "^5.5.0"
requests = "^2.32.3"
python-pptx = "^1.0.2"
python-docx = "^1.1.2"
pymupdf = "^1.24.11"
openpyxl = "^3.1.5"
cbase64 = "^0.0.9"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"
black = "^23.9.1"
isort = "^5.12.0"
pre-commit = "^3.4.0"
autopep8 = "^2.0.4"
ruff = "^0.6.9"

[tool.poetry.scripts]
db = "realitydb.rpc_server:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
exclude = '''
/(
    \.venv
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | dist
  | build
)/
'''

[tool.isort]
profile = "black"


[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = ".venv,dist,build,*.egg-info"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=realitydb --cov-report=term-missing"
